# output/final_deliverables/sprint6/api/routers/documents.py
"""
N100 Financial Intelligence Platform
Sprint 6, Day 40: Documents Endpoint

Location: src/api/routers/documents.py

documents: doc_id, company_id, report_year, annual_report
Confirmed 2026-09: no is_url_valid column exists -- spec wants this as
a returned flag, not a stored one, so it's computed here via a live
request per document.

Performance fix (2026-09): confirmed via direct testing that individual
URL checks against bseindia.com take ~1.4-1.7s each (not a timeout --
a fast, explicit HTTP 403). With TCS's 16 documents checked
SEQUENTIALLY, total endpoint time measured at 24-26 seconds -- far
outside acceptable API response time (relevant to Day 43's performance
gate). Fixed by running all checks CONCURRENTLY via a thread pool
(network-bound I/O, not CPU-bound, so threads are appropriate here)
rather than one at a time.

Accuracy note (2026-09, NOT silently worked around): confirmed both
HEAD and GET against a known-real bseindia.com PDF URL return HTTP 403
in under 2s -- not a broken link, but BSE's server rejecting the
default `requests` User-Agent (likely bot/scraper protection). A
realistic browser User-Agent header is added below since it's an
honest, general improvement (some sites genuinely gate on this), but
is_url_valid=false can still mean EITHER "the link is actually broken"
OR "this host blocks automated requests regardless of headers" -- that
distinction isn't resolvable without a full browser-based check, which
is out of scope for this endpoint. Documented here rather than implied
to be a fully solved accuracy problem.
"""

import concurrent.futures

import requests
from fastapi import APIRouter, HTTPException, Request

router = APIRouter()

# A realistic browser User-Agent -- some hosts (confirmed: bseindia.com)
# reject requests' default identifier outright with a 403, unrelated to
# whether the underlying document actually exists.
_HEADERS = {
    "User-Agent": (
        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 "
        "(KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36"
    )
}

MAX_WORKERS = 8  # concurrent checks -- network-bound, threads are fine here


def _check_url_valid(url: str, timeout: float = 5.0) -> bool:
    try:
        resp = requests.get(
            url, headers=_HEADERS, timeout=timeout, stream=True, allow_redirects=True
        )
        resp.close()
        return resp.status_code < 400
    except requests.RequestException:
        return False


@router.get("/companies/{ticker}/documents")
def get_company_documents(ticker: str, request: Request):
    """Return a company's annual report links with a concurrently-checked is_url_valid flag for each."""
    conn = request.app.state.get_db_connection()
    try:
        exists = conn.execute(
            "SELECT 1 FROM companies WHERE id = ?", (ticker,)
        ).fetchone()
        if exists is None:
            raise HTTPException(status_code=404, detail=f"Company '{ticker}' not found")

        rows = conn.execute(
            """
            SELECT doc_id, report_year, annual_report FROM documents
            WHERE company_id = ? ORDER BY report_year DESC
        """,
            (ticker,),
        ).fetchall()
    finally:
        conn.close()

    docs = [dict(row) for row in rows]

    # Check all document URLs concurrently instead of one at a time --
    # confirmed 2026-09: sequential checks took 24-26s for 16 documents.
    if docs:
        with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_WORKERS) as executor:
            future_to_doc = {
                executor.submit(_check_url_valid, d["annual_report"]): d for d in docs
            }
            for future in concurrent.futures.as_completed(future_to_doc):
                doc = future_to_doc[future]
                doc["is_url_valid"] = future.result()

    return {"company_id": ticker, "count": len(docs), "results": docs}